requirements.txt
openai
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
psutil==5.9.6
pydantic-settings==2.1.0